This PR unifies `utilities.py`.
