--- notebooks/HF_radar/utilities.py
@@ -1,351 +1,0 @@
-import fnmatch
-import lxml.html
-from urllib import urlopen
-
-import requests
-
-# Scientific stack.
-import iris
-iris.FUTURE.netcdf_promote = True
-from iris.cube import CubeList
-from iris.exceptions import CoordinateNotFoundError, CoordinateMultiDimError
-
-import numpy as np
-import numpy.ma as ma
-
-from pandas import DataFrame
-from netCDF4 import Dataset, MFDataset, date2index, num2date
-from IPython.display import HTML
-
-from owslib import fes
-
-
-def find_timevar(cube):
-    """Return the time variable from Iris."""
-    try:
-        cube.coord(axis='T').rename('time')
-    except CoordinateNotFoundError:
-        pass
-    timevar = cube.coord('time')
-    return timevar
-
-
-def time_near(cube, start):
-    """Return the nearest time to `start`."""
-    timevar = find_timevar(cube)
-    try:
-        time = timevar.units.date2num(start)
-        itime = timevar.nearest_neighbour_index(time)
-    except IndexError:
-        itime = -1
-    return timevar.points[itime]
-
-
-def slice_bbox_extract(cube, bbox):
-    """Extract a subsetted cube inside a lon,lat bounding box
-    bbox = [[lon_min, lat_min], [lon_max, lat_max]]."""
-    lons = cube.coord('longitude').points
-    lats = cube.coord('latitude').points
-
-    def minmax(v):
-        return np.min(v), np.max(v)
-    inregion = np.logical_and(np.logical_and(lons > bbox[0][0],
-                                             lons < bbox[1][0]),
-                              np.logical_and(lats > bbox[1][0],
-                                             lats < bbox[1][1]))
-    region_inds = np.where(inregion)
-    imin, imax = minmax(region_inds[0])
-    jmin, jmax = minmax(region_inds[1])
-    return cube[..., imin:imax+1, jmin:jmax+1]
-
-
-def intersection(cube, bbox):
-    """Sub sets cube with 1D or 2D lon, lat coords.
-    Using `intersection` instead of `extract` we deal with 0-360
-    longitudes automagically."""
-    try:
-        cube = cube.intersection(longitude=(bbox[0][0], bbox[1][0]),
-                                 latitude=(bbox[0][1], bbox[1][1]))
-    except CoordinateMultiDimError:
-        cube = slice_bbox_extract(cube, bbox)
-    return cube
-
-
-def get_cubes(url, name_list=None, bbox=None, time=None):
-    """Load and time/space-slice the cubes from a given `url`.
-    name_list : CF standard_name list of cubes
-    bbox : [[lon_min, lat_min], [lon_max, lat_max]]
-    time : datetime object (TODO: make it a time range.)
-    """
-    cubes = iris.load_raw(url)
-    if name_list:
-        in_list = lambda cube: cube.standard_name in name_list
-        cubes = CubeList([cube for cube in cubes if in_list(cube)])
-        cubes = cubes.merge()
-    if time:
-        cubes = cubes.extract(iris.Constraint(time=time_near(cubes[0], time)))
-    if bbox:
-        cubes = CubeList([intersection(cube, bbox) for cube in cubes])
-    return cubes
-
-
-def rot2d(x, y, ang):
-    """Rotate vectors by geometric angle."""
-    xr = x * np.cos(ang) - y * np.sin(ang)
-    yr = x * np.sin(ang) + y * np.cos(ang)
-    return xr, yr
-
-
-def shrink(a, b):
-    """Return array shrunk to fit a specified shape by trimming or averaging.
-
-    a = shrink(array, shape)
-
-    array is an numpy ndarray, and shape is a tuple (e.g., from
-    array.shape).  `a` is the input array shrunk such that its maximum
-    dimensions are given by shape. If shape has more dimensions than
-    array, the last dimensions of shape are fit.
-
-    as, bs = shrink(a, b)
-
-    If the second argument is also an array, both a and b are shrunk to
-    the dimensions of each other. The input arrays must have the same
-    number of dimensions, and the resulting arrays will have the same
-    shape.
-    Example
-    -------
-
-    >>> shrink(rand(10, 10), (5, 9, 18)).shape
-    (9, 10)
-    >>> map(shape, shrink(rand(10, 10, 10), rand(5, 9, 18)))
-    [(5, 9, 10), (5, 9, 10)]
-
-    """
-
-    if isinstance(b, np.ndarray):
-        if not len(a.shape) == len(b.shape):
-            raise Exception('Input arrays must have the same number of'
-                            'dimensions')
-        a = shrink(a, b.shape)
-        b = shrink(b, a.shape)
-        return (a, b)
-
-    if isinstance(b, int):
-        b = (b,)
-
-    if len(a.shape) == 1:  # 1D array is a special case
-        dim = b[-1]
-        while a.shape[0] > dim:  # Only shrink a.
-            if (dim - a.shape[0]) >= 2:  # Trim off edges evenly.
-                a = a[1:-1]
-            else:  # Or average adjacent cells.
-                a = 0.5*(a[1:] + a[:-1])
-    else:
-        for dim_idx in range(-(len(a.shape)), 0):
-            dim = b[dim_idx]
-            a = a.swapaxes(0, dim_idx)  # Put working dim first
-            while a.shape[0] > dim:  # Only shrink a
-                if (a.shape[0] - dim) >= 2:  # trim off edges evenly
-                    a = a[1:-1, :]
-                if (a.shape[0] - dim) == 1:  # Or average adjacent cells.
-                    a = 0.5*(a[1:, :] + a[:-1, :])
-            a = a.swapaxes(0, dim_idx)  # Swap working dim back.
-    return a
-
-
-def get_roms(url, time_slice, n=3):
-    with Dataset(url) as nc:
-        ncv = nc.variables
-        time = ncv['ocean_time']
-        tidx = date2index(time_slice, time, select='nearest')
-        time = num2date(time[tidx], time.units, time.calendar)
-
-        mask = ncv['mask_rho'][:]
-        lon_rho = ncv['lon_rho'][:]
-        lat_rho = ncv['lat_rho'][:]
-        anglev = ncv['angle'][:]
-
-        u = ncv['u'][tidx, -1, ...]
-        v = ncv['v'][tidx, -1, ...]
-
-        u = shrink(u, mask[1:-1, 1:-1].shape)
-        v = shrink(v, mask[1:-1, 1:-1].shape)
-
-        u, v = rot2d(u, v, anglev[1:-1, 1:-1])
-
-        lon = lon_rho[1:-1, 1:-1]
-        lat = lat_rho[1:-1, 1:-1]
-
-        u, v = u[::n, ::n], v[::n, ::n]
-        lon, lat = lon[::n, ::n], lat[::n, ::n]
-
-        u = ma.masked_invalid(u)
-        v = ma.masked_invalid(v)
-    return dict(lon=lon, lat=lat, u=u, v=v, time=time)
-
-
-def date_range(start_date='1900-01-01', stop_date='2100-01-01',
-               constraint='overlaps'):
-    """Hopefully something like this will be implemented in fes soon."""
-    if constraint == 'overlaps':
-        propertyname = 'apiso:TempExtent_begin'
-        start = fes.PropertyIsLessThanOrEqualTo(propertyname=propertyname,
-                                                literal=stop_date)
-        propertyname = 'apiso:TempExtent_end'
-        stop = fes.PropertyIsGreaterThanOrEqualTo(propertyname=propertyname,
-                                                  literal=start_date)
-    elif constraint == 'within':
-        propertyname = 'apiso:TempExtent_begin'
-        start = fes.PropertyIsGreaterThanOrEqualTo(propertyname=propertyname,
-                                                   literal=start_date)
-        propertyname = 'apiso:TempExtent_end'
-        stop = fes.PropertyIsLessThanOrEqualTo(propertyname=propertyname,
-                                               literal=stop_date)
-    return start, stop
-
-
-def service_urls(records, service='odp:url'):
-    """Extract service_urls of a specific type (DAP, SOS) from records."""
-    service_string = 'urn:x-esri:specification:ServiceType:' + service
-    urls = []
-    for key, rec in records.items():
-        # Create a generator object, and iterate through it until the match is
-        # found if not found, gets the default value (here "none").
-        url = next((d['url'] for d in rec.references if
-                    d['scheme'] == service_string), None)
-        if url is not None:
-            urls.append(url)
-    return urls
-
-
-def get_coordinates(bounding_box, bounding_box_type=''):
-    """Create bounding box coordinates for the map."""
-    coordinates = []
-    if bounding_box_type == "box":
-        coordinates.append([bounding_box[1], bounding_box[0]])
-        coordinates.append([bounding_box[1], bounding_box[2]])
-        coordinates.append([bounding_box[3], bounding_box[2]])
-        coordinates.append([bounding_box[3], bounding_box[0]])
-        coordinates.append([bounding_box[1], bounding_box[0]])
-    return coordinates
-
-
-def inline_map(m):
-    """From http://nbviewer.ipython.org/gist/rsignell-usgs/
-    bea6c0fe00a7d6e3249c."""
-    m._build_map()
-    srcdoc = m.HTML.replace('"', '&quot;')
-    embed = HTML('<iframe srcdoc="{srcdoc}" '
-                 'style="width: 100%; height: 500px; '
-                 'border: none"></iframe>'.format(srcdoc=srcdoc))
-    return embed
-
-
-def css_styles():
-    style = """
-        <style>
-        .info {
-            background-color: #fcf8e3; border-color: #faebcc;
-                border-left: 5px solid #8a6d3b; padding: 0.5em; color: #8a6d3b;
-        }
-        .success {
-            background-color: #d9edf7; border-color: #bce8f1;
-                border-left: 5px solid #31708f; padding: 0.5em; color: #31708f;
-        }
-        .error {
-            background-color: #f2dede; border-color: #ebccd1;
-                border-left: 5px solid #a94442; padding: 0.5em; color: #a94442;
-        }
-        .warning {
-            background-color: #fcf8e3; border-color: #faebcc;
-                border-left: 5px solid #8a6d3b; padding: 0.5em; color: #8a6d3b;
-        }
-        .text-shadow {
-            text-shadow: 0 1px 0 #ccc,0 2px 0 #c9c9c9,0 3px 0 #bbb,
-            0 4px 0 #b9b9b9,0 5px 0 #aaa,0 6px 1px rgba(0,0,0,.1)
-            }
-        </style>
-    """
-    return HTML(style)
-
-
-def processStationInfo(obs_loc_df, st_list, source):
-    st_data = obs_loc_df['station_id']
-    lat_data = obs_loc_df['latitude (degree)']
-    lon_data = obs_loc_df['longitude (degree)']
-    for k in range(0, len(st_data)):
-        station_name = st_data[k]
-        if station_name in st_list:
-            pass
-        else:
-            st_list[station_name] = {}
-            st_list[station_name]["lat"] = lat_data[k]
-            st_list[station_name]["source"] = source
-            st_list[station_name]["lon"] = lon_data[k]
-            print(station_name)
-    print("number of stations in bbox %s" % len(st_list.keys()))
-    return st_list
-
-
-def get_ncfiles_catalog(station_id, jd_start, jd_stop):
-    station_name = station_id.split(":")[-1]
-    uri = 'http://dods.ndbc.noaa.gov/thredds/dodsC/data/adcp'
-    url = ('%s/%s/' % (uri, station_name))
-    urls = _url_lister(url)
-    filetype = "*.nc"
-    file_list = [filename for filename in fnmatch.filter(urls, filetype)]
-    files = [fname.split('/')[-1] for fname in file_list]
-    urls = ['%s/%s/%s' % (uri, station_name, fname) for fname in files]
-
-    nc = MFDataset(urls)
-
-    time_dim = nc.variables['time']
-    calendar = 'gregorian'
-    idx_start = date2index(jd_start, time_dim, calendar=calendar,
-                           select='nearest')
-    idx_stop = date2index(jd_stop, time_dim, calendar=calendar,
-                          select='nearest')
-
-    dir_dim = nc.variables['water_dir'][idx_start:idx_stop, ...].squeeze()
-    speed_dim = nc.variables['water_spd'][idx_start:idx_stop, ...].squeeze()
-    if dir_dim.ndim != 1:
-        dir_dim = dir_dim[:, 0]
-        speed_dim = speed_dim[:, 0]
-    time_dim = nc.variables['time']
-    dates = num2date(time_dim[idx_start:idx_stop],
-                     units=time_dim.units,
-                     calendar='gregorian').squeeze()
-    data = dict()
-    data['sea_water_speed (cm/s)'] = speed_dim
-    col = 'direction_of_sea_water_velocity (degree)'
-    data[col] = dir_dim
-    time = dates
-    columns = ['sea_water_speed (cm/s)',
-               'direction_of_sea_water_velocity (degree)']
-    df = DataFrame(data=data, index=time, columns=columns)
-    return df
-
-
-def _url_lister(url):
-    urls = []
-    connection = urlopen(url)
-    dom = lxml.html.fromstring(connection.read())
-    for link in dom.xpath('//a/@href'):
-        urls.append(link)
-    return urls
-
-
-def sos_request(url='http://opendap.co-ops.nos.noaa.gov/ioos-dif-sos/SOS', **kw):
-    offering = 'urn:ioos:network:NOAA.NOS.CO-OPS:CurrentsActive'
-    params = dict(service='SOS',
-                  request='GetObservation',
-                  version='1.0.0', bin='1',
-                  offering=offering,
-                  responseFormat='text/csv')
-    params.update(kw)
-    r = requests.get(url, params=params)
-    r.raise_for_status()
-    if 'excel' in r.headers['Content-Type'] or 'csv' in r.headers['Content-Type']:
-        return r.url
-    else:
-        raise NameError('Bad url {}'.format(r.url))
